--- frontend/src/media/tiny.csv
@@ -1,11 +1,0 @@
-name,value
-31-Aug-16 1:00:00 PM CDT,512849984.00 BTU
-31-Aug-16 1:15:04 PM CDT,512849984.00 BTU
-31-Aug-16 1:30:04 PM CDT,512849984.00 BTU
-31-Aug-16 1:45:04 PM CDT,512849984.00 BTU
-31-Aug-16 2:00:04 PM CDT,512889984.00 BTU
-31-Aug-16 2:15:04 PM CDT,512889984.00 BTU
-31-Aug-16 2:30:04 PM CDT,512889984.00 BTU
-31-Aug-16 2:45:04 PM CDT,512900000.00 BTU
-31-Aug-16 3:00:04 PM CDT,512960000.00 BTU
-31-Aug-16 3:15:04 PM CDT,512990016.00 BTU
